% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{parse_odds_url}
\alias{parse_odds_url}
\title{parse_odds_url}
\usage{
parse_odds_url(url)
}
\arguments{
\item{url}{A URL reference to the odds endpoint of the Action Network API.}
}
\value{
The odds data for the game, returned in rectangular format.
}
\description{
A helper function to generically parse data returned from the Action Network odds API.
}
