% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{scrape_cfb_game_odds}
\alias{scrape_cfb_game_odds}
\title{scrape_cfb_game_odds}
\usage{
scrape_cfb_game_odds(
  week,
  season_type,
  season,
  division = "FBS",
  period = "event"
)
}
\arguments{
\item{week}{The day to obtain data from. Ranges from 0-4 for the preseason,
from 0-15 for the regular season (including conference championships), and
just 1 for the postseason.}

\item{season_type}{The type of season to obtain data for. Must be one of
"reg", or "post".}

\item{season}{The season to obtain data from.}

\item{division}{One of "FBS","FCS", or "D2".}

\item{period}{One of "event", "firsthalf", "secondhalf", "firstquarter", "secondquarter", "thirdquarter", or "fourthquarter". Defaults to "event".}
}
\value{
a dataframe of odds from the Action Network.
}
\description{
Obtains historical and future CFB odds from the Action Network
}
