% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.R
\name{scrape_nba_game_odds}
\alias{scrape_nba_game_odds}
\title{scrape_nba_game_odds}
\usage{
scrape_nba_game_odds(date = Sys.Date())
}
\arguments{
\item{date}{the day to scrape odds from. Defaults to today}
}
\value{
a dataframe of odds from the Action Network.
}
\description{
Obtains historical and future NBA odds from the Action Network
}
