% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl.R
\name{scrape_nfl_game_odds}
\alias{scrape_nfl_game_odds}
\title{scrape_nfl_game_odds}
\usage{
scrape_nfl_game_odds(week, season_type, season)
}
\arguments{
\item{week}{The week to obtain data from. Ranges from 0-4 for the preseason,
from 0-18 for the regular season, and from 0-5 for the postseason.}

\item{season_type}{The type of season to obtain data for. Must be one of
"pre", "reg", or "post".}

\item{season}{The season to obtain data from.}
}
\value{
a dataframe of odds from the Action Network.
}
\description{
Obtains historical and future NFL odds from the Action Network
}
