% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{scrape_wnba_game_odds}
\alias{scrape_wnba_game_odds}
\title{scrape_wnba_game_odds}
\usage{
scrape_wnba_game_odds(date = Sys.Date(), period = "event")
}
\arguments{
\item{period}{}
}
\description{
scrape_wnba_game_odds
}
