% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{scrape_ncaab_game_odds}
\alias{scrape_ncaab_game_odds}
\title{scrape_ncaab_game_odds}
\usage{
scrape_ncaab_game_odds(date = Sys.Date(), period = "event")
}
\arguments{
\item{date}{The date to obtain odds for. Defaults to the current day.}

\item{period}{One of "event", "firsthalf", "secondhalf". Defaults to "event".}
}
\value{
A dataframe of odds for men's NCAA basketball games on the specified date.
}
\description{
Return the odds for a given period for all games on a specified date for NCAA men's basketball games from the Action Network.
}
