% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{scrape_wncaab_game_odds}
\alias{scrape_wncaab_game_odds}
\title{scrape_wncaab_game_odds}
\usage{
scrape_wncaab_game_odds(date = Sys.Date())
}
\arguments{
\item{date}{}
}
\description{
scrape_wncaab_game_odds
}
